修改砍价入参数据floor_price错误问题

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/market/bargain/BargainAddParam.java
@@ -77,7 +77,7 @@ public class BargainAddParam {
 	/**
 	 * 任意金额结算模式的结算金额底价 
 	 */
-	private Double floorPrice;
+	private BigDecimal floorPrice;
 	
 	/**
 	 * 活动商品库存
